to int

+++ gordo/machine/dataset/data_provider/providers.py
@@ -384,7 +384,7 @@ class RandomDataProvider(GordoBaseDataProvider):
             #nr = random.randint(self.min_size, self.max_size)
 
             # ensure a observation for each aggregated time period (using default 10T)
-            nr = (pd.Timestamp(train_end_date) - pd.Timestamp(train_start_date)) / pd.Timedelta("1 hour") * 6
+            nr = int((pd.Timestamp(train_end_date) - pd.Timestamp(train_start_date)) / pd.Timedelta("1 hour") * 6)
 
             random_index = self._random_dates(train_start_date, train_end_date, n=nr)
             series = pd.Series(
